Fix/cp ajustes varios filtros cache

+++ FrizzApp/FirzzApp.Business/Dtos/RequestDto/GetAllProductDto.cs
@@ -9,8 +9,8 @@ namespace FirzzApp.Business.Dtos.RequestDto
         public int NumeroPagina { get; set; }
         public int CantidadPagina { get; set; }
         public int? CategoriaId { get; set; }
-        public int? PrecioMinimo { get; set; }
-        public int? PrecioMaximo { get; set; }
+        public decimal? PrecioMinimo { get; set; }
+        public decimal? PrecioMaximo { get; set; }
         public CacheTypeEnum CacheType { get; set; }
     }
 }
